src/requirements.txt
about-time==4.2.1
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
alive-progress==3.1.4
annotated-types==0.5.0
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
black==23.9.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
frozenlist==1.3.3
gevent==23.9.0.post1
grapheme==0.6.0
greenlet==2.0.2
grequests==0.7.0
humanize==4.8.0
idna==3.4
importlib-metadata==6.7.0
isort==5.12.0
magic-filter==1.0.11
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.25.2
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
platformdirs==3.10.0
postgres==4.0
progress==1.6
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
